+++ src/data-source/data-source.test.ts
@@ -217,7 +217,7 @@ describe('DataSource', () => {
    */
   describe('applyTemplateVariables', () => {
     type KeyType = keyof RedisQuery;
-    const testedFieldKeys: KeyType[] = ['keyName', 'query', 'field', 'filter', 'legend', 'value'];
+    const testedFieldKeys: KeyType[] = ['keyName', 'query', 'field', 'filter', 'legend', 'value', 'path', 'cypher'];
 
     testedFieldKeys.forEach((fieldKey) => {
       describe(fieldKey, () => {