+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Disable Command-line interface in the Query Editor (#226)
 - Support of ZRANGE command (#182)
 - Upgrade Grafana 8.2.3 and backend dependencies
+- Support fetching from RedisJSON datasource (JSON.GET, JSON.TYPE, JSON.ARRLEN, JSON.OBJLEN, JSON.OBJKEYS)
 
 ### Bug fixes